Move simple sender to own file
